*** empty log message ***

+++ src/rxvt.h
@@ -1,5 +1,5 @@
 /*
- * $Id: rxvt.h,v 1.19 2003/12/19 09:41:09 pcg Exp $
+ * $Id: rxvt.h,v 1.20 2003/12/24 09:22:29 pcg Exp $
  */
 
 #ifndef _RXVT_H_                /* include once only */
@@ -887,7 +887,7 @@ struct rxvt_term : rxvt_vars {
 #endif
   short           rvideo;
   int16_t         num_scr;    /* screen: number lines scrolled             */
-  uint16_t        prev_ncol,   /* screen: previous number of columns        */
+  unsigned int    prev_ncol,  /* screen: previous number of columns        */
                   prev_nrow;  /* screen: previous number of rows           */
 #ifdef RXVT_GRAPHICS
   uint16_t        gr_prev_start;
